Break the plugin examples in part.
It fixes circular dependencies when entrypoints are loaded
