+++ docs/source/plugins.rst
@@ -98,19 +98,19 @@ The example package contains a ``setup.py`` that declares a number of
         ...
         entry_points={
             # Define a Tracking Store plugin for tracking URIs with scheme 'file-plugin'
-            "mlflow.tracking_store": "file-plugin=mlflow_test_plugin:PluginFileStore",
+            "mlflow.tracking_store": "file-plugin=mlflow_test_plugin.file_store:PluginFileStore",
             # Define a ArtifactRepository plugin for artifact URIs with scheme 'file-plugin'
             "mlflow.artifact_repository":
-                "file-plugin=mlflow_test_plugin:PluginLocalArtifactRepository",
+                "file-plugin=mlflow_test_plugin.local_artifact:PluginLocalArtifactRepository",
             # Define a RunContextProvider plugin. The entry point name for run context providers
             # is not used, and so is set to the string "unused" here
-            "mlflow.run_context_provider": "unused=mlflow_test_plugin:PluginRunContextProvider",
+            "mlflow.run_context_provider": "unused=mlflow_test_plugin.run_context_provider:PluginRunContextProvider",
             # Define a Model Registry Store plugin for tracking URIs with scheme 'file-plugin'
             "mlflow.model_registry_store":
-                "file-plugin=mlflow_test_plugin:PluginRegistrySqlAlchemyStore",
+                "file-plugin=mlflow_test_plugin.sqlalchemy_store:PluginRegistrySqlAlchemyStore",
             # Define a MLflow Project Backend plugin called 'dummy-backend'
             "mlflow.mlproject_backend":
-                "dummy-backend=mlflow_test_plugin:PluginDummyProjectBackend",
+                "dummy-backend=mlflow_test_plugin.dummy_backend:PluginDummyProjectBackend",
         },
     )
 
@@ -131,7 +131,7 @@ plugin:
    * - Plugins for overriding definitions of tracking APIs like ``mlflow.log_metric``, ``mlflow.start_run`` for a specific
        tracking URI scheme.
      - mlflow.tracking_store
-     - The entry point value (e.g. ``mlflow_test_plugin:PluginFileStore``) specifies a custom subclass of
+     - The entry point value (e.g. ``mlflow_test_plugin.local_store:PluginFileStore``) specifies a custom subclass of
        `mlflow.tracking.store.AbstractStore <https://github.com/mlflow/mlflow/blob/branch-1.5/mlflow/store/tracking/abstract_store.py#L8>`_
        (e.g., the `PluginFileStore class <https://github.com/mlflow/mlflow/blob/branch-1.5/tests/resources/mlflow-test-plugin/mlflow_test_plugin/__init__.py#L9>`_
        within the ``mlflow_test_plugin`` module).
@@ -145,7 +145,7 @@ plugin:
    * - Plugins for defining artifact read/write APIs like ``mlflow.log_artifact``, ``MlflowClient.download_artifacts`` for a specified
        artifact URI scheme (e.g. the scheme used by your in-house blob storage system).
      - mlflow.artifact_repository
-     - The entry point value (e.g. ``mlflow_test_plugin:PluginLocalArtifactRepository``) specifies a custom subclass of
+     - The entry point value (e.g. ``mlflow_test_plugin.local_artifact:PluginLocalArtifactRepository``) specifies a custom subclass of
        `mlflow.store.artifact.artifact_repo.ArtifactRepository <https://github.com/mlflow/mlflow/blob/branch-1.5/mlflow/store/artifact/artifact_repo.py#L12>`_
        (e.g., the `PluginLocalArtifactRepository class <https://github.com/mlflow/mlflow/blob/branch-1.5/tests/resources/mlflow-test-plugin/mlflow_test_plugin/__init__.py#L18>`_
        within the ``mlflow_test_plugin`` module).
@@ -160,7 +160,7 @@ plugin:
 
    * - Plugins for specifying custom context tags at run creation time, e.g. tags identifying the git repository associated with a run.
      - mlflow.run_context_provider
-     - The entry point name is unused. The entry point value (e.g. ``mlflow_test_plugin:PluginRunContextProvider``) specifies a custom subclass of
+     - The entry point name is unused. The entry point value (e.g. ``mlflow_test_plugin.run_context_provider:PluginRunContextProvider``) specifies a custom subclass of
        `mlflow.tracking.context.abstract_context.RunContextProvider <https://github.com/mlflow/mlflow/blob/branch-1.5/mlflow/tracking/context/abstract_context.py#L4>`_
        (e.g., the `PluginRunContextProvider class <https://github.com/mlflow/mlflow/blob/branch-1.5/tests/resources/mlflow-test-plugin/mlflow_test_plugin/__init__.py#L23>`_
        within the ``mlflow_test_plugin`` module) to register.
@@ -171,7 +171,7 @@ plugin:
      - mlflow.model_registry_store
      - .. note:: The Model Registry is in beta (as of MLflow 1.5). Model Registry APIs are not guaranteed to be stable, and Model Registry plugins may break in the future.
 
-       The entry point value (e.g. ``mlflow_test_plugin:PluginRegistrySqlAlchemyStore``) specifies a custom subclass of
+       The entry point value (e.g. ``mlflow_test_plugin.sqlalchemy_store:PluginRegistrySqlAlchemyStore``) specifies a custom subclass of
        `mlflow.tracking.model_registry.AbstractStore <https://github.com/mlflow/mlflow/blob/branch-1.5/mlflow/store/model_registry/abstract_store.py#L6>`_
        (e.g., the `PluginRegistrySqlAlchemyStore class <https://github.com/mlflow/mlflow/blob/branch-1.5/tests/resources/mlflow-test-plugin/mlflow_test_plugin/__init__.py#L33>`_
        within the ``mlflow_test_plugin`` module)
@@ -184,7 +184,7 @@ plugin:
 
    * - Plugins for overridding MLflow Project Execution backend like databricks or kubernetes
      - mlflow.project.backend
-     - The entry point value (e.g. ``mlflow_test_plugin:PluginDummyProjectBackend``) specifies a custom subclass of
+     - The entry point value (e.g. ``mlflow_test_plugin.dummy_backend:PluginDummyProjectBackend``) specifies a custom subclass of
        ``mlflow.project.backend.AbstractBackend``)
      - `DatabricksProjectBackend <https://github.com/mlflow/mlflow/blob/master/mlflow/projects/backend/databricks.py#L291>`_
