+++ tests/resources/mlflow-test-plugin/mlflow_test_plugin/dummy_backend.py
@@ -1,0 +1,35 @@
+from mlflow.entities import RunStatus
+from mlflow.projects.backend.abstract_backend import AbstractBackend
+from mlflow.projects.submitted_run import SubmittedRun
+
+
+class DummySubmittedRun(SubmittedRun):
+    """
+    A run that just does nothing
+    """
+
+    def __init__(self, run_id):
+        self._run_id = run_id
+
+    def wait(self):
+        return True
+
+    def get_status(self):
+        return RunStatus.FINISHED
+
+    def cancel(self):
+        pass
+
+    @property
+    def run_id(self):
+        return self._run_id
+
+
+class PluginDummyProjectBackend(AbstractBackend):
+    def run(self, active_run, uri, entry_point, parameters,
+            backend_config):
+        return DummySubmittedRun(active_run)
+
+    @property
+    def name(self):
+        return "dummy"
