+++ tests/resources/mlflow-test-plugin/mlflow_test_plugin/local_artifact.py
@@ -1,0 +1,6 @@
+from mlflow.store.artifact.local_artifact_repo import LocalArtifactRepository
+
+
+class PluginLocalArtifactRepository(LocalArtifactRepository):
+    """LocalArtifactRepository provided through plugin system"""
+    is_plugin = True
